Remove batch answer updates, wait before redirecting, refactor helpers
